+++ packages/shared/src/index.ts
@@ -5,6 +5,7 @@
  */
 export * from "./locales";
 export * from "./permissions";
+export * as mediaSchemas from "./schemas/media";
 export * as propertySchemas from "./schemas/property";
 export * as uploadSchemas from "./schemas/upload";